Get transcript biotypes

List all available transcript biotypes in the EnsDb.Hsapiens.v75 database.

Usage

get_tx_biotypes(tx_biotypes = NULL, as_filter = TRUE, verbose = TRUE)

Arguments

tx_biotypes

Transcript biotypes to include in the gene model track. By default (NULL), all transcript biotypes will be included. See get_tx_biotypes for a full list of all available biotypes

as_filter

Return results as an TxBiotypeFilter (default: TRUE). Otherwise, a character vector will be returned.

verbose

Print messages.

Value

Character vector of valid transcript biotypes.

Examples

tx_filter <- get_tx_biotypes()